ref(memory): remove store facade
