+++ packages/junior-memory/src/agent.ts
@@ -9,7 +9,7 @@ import {
   memorySupersessionInputSchema,
   type MemorySupersessionDecision,
   type MemorySupersessionInput,
-} from "./store";
+} from "./create";
 import {
   MEMORY_KINDS,
   memoryRuntimeContextSchema,
