+++ packages/junior-memory/src/index.ts
@@ -8,18 +8,5 @@ export {
   type MemoryListResponse,
 } from "./api";
 export type { MemoryPluginOptions } from "./plugin";
-export { createMemoryStore } from "./store";
-export type {
-  ArchiveMemoryInput,
-  CreateMemoryInput,
-  CreateMemoryResult,
-  ListMemoriesInput,
-  MemoryDb,
-  MemoryEmbeddingProvider,
-  MemoryRecord,
-  MemoryStore,
-  MemoryStoreOptions,
-  SearchMemoriesInput,
-} from "./store";
 export { MEMORY_KINDS } from "./types";
 export type { MemoryKind, MemoryRuntimeContext } from "./types";
